+++ CHANGELOG.md
@@ -13,7 +13,7 @@ All notable changes to this project will be documented in this file.
 - [Chore]: remove kgpAndroidxComposeCompiler
 - [Chore]: update dokka to 2.0.0
 - [Chore]: update Gradle to 8.13
-- [Chore]: update AGP to 8.9.1 and make new minimum. Minimum supported Gradle version is now 8.11.1
+- [Chore]: update AGP to 8.10.0 and make new minimum 8.9.0. Minimum supported Gradle version is now 8.11.1
 - [Chore]: update dagger to 2.56.2
 - [Chore]: update android junit 5 plugin to 1.12.0.0. JUnit BOM to 5.12.0
 - [Chore]: update kotlinter to 5.0.2