+++ CHANGELOG.md
@@ -18,7 +18,7 @@ All notable changes to this project will be documented in this file.
 - [Chore]: update dagger to 2.56.2
 - [Chore]: update android junit 5 plugin to 1.12.0.0. JUnit BOM to 5.12.0
 - [Chore]: update kotlinter to 5.0.2
-- [Chore]: update kover to 0.9.1
+- [Chore]: update kover to 0.9.9
 - [Chore]: update ksp to 2.1.20-1.0.32
 - [Chore]: update maven publish plugin to 0.52.0
 - [Chore]: update room to 2.7.0